refactor: server translation locale fallbacks
