+++ components/page/HeaderProgress/styles.ts
@@ -1,4 +1,8 @@
 import styled, { css } from "styled-components";
+import {
+  Marker,
+  ProgressBar as BaseProgressBar,
+} from "@rubin-epo/epo-react-lib";
 
 const colors = [
   "#019305",
@@ -13,10 +17,10 @@ export const HeaderProgress = styled.ul<{
   $padding: boolean;
 }>`
   display: flex;
-  gap: calc(var(--PADDING-SMALL, 20px) / 2);
   align-items: center;
   width: 100%;
   height: 30px;
+  gap: calc(var(--PADDING-SMALL, 20px) / 2);
   ${({ $padding }) =>
     $padding
       ? css`
@@ -25,18 +29,22 @@ export const HeaderProgress = styled.ul<{
       : ""}
   background-color: ${({ $backgroundColor }) => $backgroundColor}
 `;
+export const ProgressBar = styled(BaseProgressBar)``;
 
 export const SectionProgress = styled.li.attrs<{ $proportion: number }>(
   ({ $proportion }) => ({ style: { flexBasis: `${$proportion}%` } })
 )<{ $proportion: number }>`
   list-style-type: none;
-  ${colors.reduce((prev, curr, i) =>
-    prev.concat(`
+  ${colors.reduce(
+    (prev, curr, i) =>
+      prev.concat(`
       &:nth-of-type(${colors.length}n + ${i + 1}) {
         ${ProgressBar} {
           --progress-bar-background: ${curr}
         }
       }
-    `)
+    `),
+    css``
   )}
 `;
+export const IconMarker = styled(Marker)<{ $backgroundColor: string }>``;
