+++ components/templates/HomePage/index.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import { graphql, useFragment, FragmentType } from "@/gql/public-schema";
 import ContentBlockFactory from "@/components/factories/ContentBlockFactory";
 import * as Styled from "./styles";
