+++ lib/i18n/index.ts
@@ -1,7 +1,4 @@
-import { createInstance } from "i18next";
 import resourcesToBackend from "i18next-resources-to-backend";
-import { initReactI18next } from "react-i18next/initReactI18next";
-import { getOptions } from "./settings";
 
 export const loadResources = resourcesToBackend(
   (language: string, namespace: string, callback) => {
@@ -40,32 +37,3 @@ export const loadResources = resourcesToBackend(
     }
   }
 );
-
-const initI18next = async (lng: string, ns: string | string[]) => {
-  // on server side we create a new instance for each render, because during compilation everything seems to be executed in parallel
-  const i18nInstance = createInstance();
-  await i18nInstance
-    .use(initReactI18next)
-    .use(loadResources)
-    .init(getOptions(lng, ns));
-  return i18nInstance;
-};
-
-async function useTranslation(
-  lng: string,
-  ns: string | string[],
-  options: any = {}
-) {
-  const i18nextInstance = await initI18next(lng, ns);
-
-  return {
-    t: i18nextInstance.getFixedT(
-      lng,
-      Array.isArray(ns) ? ns[0] : ns,
-      options.keyPrefix
-    ),
-    i18n: i18nextInstance,
-  };
-}
-
-export { useTranslation, useTranslation as serverTranslation };
