+++ package.json
@@ -82,7 +82,7 @@
     "i18next-resources-to-backend": "^1.2.1",
     "jwt-decode": "^4.0.0",
     "lodash": "^4.17.21",
-    "next": "^14.2.10",
+    "next": "^14.2.25",
     "next-build-id": "^3.0.0",
     "next-intl": "^3.0.3",
     "npm-run-all": "^4.1.5",
@@ -94,6 +94,7 @@
     "react-i18next": "13",
     "react-is": "^18.2.0",
     "screenfull": "^6.0.2",
+    "server-only": "^0.0.1",
     "styled-components": "^6.1.1",
     "swr": "^2.2.4",
     "temml": "^0.10.23",
